# Conflicts:
#	Tables/Sources/gameplay/MainTask/DailyTask.xlsx
#	Tables/Sources/gameplay/MainTask/MainTask.xlsx

+++ UnityProject/SimSwordsman/Assets/Scripts/Game/UIScripts/GamePanels/FacilityPanel/ForgeHousePanel/ForgeHousePanel.cs
@@ -46,7 +46,7 @@ namespace GameWish.Game
             m_UpgradeBtn.onClick.AddListener(() =>
             {
 
-                bool isReduceSuccess = GameDataMgr.S.GetPlayerData().ReduceCoinNum(double.Parse(m_UpgradeRequiredCoinTxt.text));
+                bool isReduceSuccess = GameDataMgr.S.GetPlayerData().ReduceCoinNum(long.Parse(m_UpgradeRequiredCoinTxt.text));
 
                 if (isReduceSuccess)
                 {
